+++ src/views/DestinationDetailView.vue
@@ -14,6 +14,7 @@ import PackingChecklist from '../components/PackingChecklist.vue';
 import TravelMapVisualizer from '../components/TravelMapVisualizer.vue';
 import WeatherClimateWidget from '../components/WeatherClimateWidget.vue';
 import UserTravelStories from '../components/UserTravelStories.vue';
+import AudioGuidePlayer from '../components/AudioGuidePlayer.vue';
 
 const route = useRoute();
 const travelStore = useTravelStore();
@@ -54,6 +55,7 @@ const relatedPosts = computed(() => {
           </ul>
         </div>
 
+        <AudioGuidePlayer :destination-id="destination.id" class="margin-top-calc" />
         <WeatherClimateWidget :destination-id="destination.id" class="margin-top-calc" />
         <BudgetCalculator :destination-id="destination.id" class="margin-top-calc" />
         <ItineraryBuilder :destination-id="destination.id" class="margin-top-calc" />